+++ src/main/java/com/Rpg/sistem_mayfair/repository/FamiliaRepository.java
@@ -3,5 +3,5 @@ package com.Rpg.sistem_mayfair.repository;
 import com.Rpg.sistem_mayfair.domain.Familia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FamiliaRepository extends JpaRepository<Familia, Integer> {
+public interface FamiliaRepository extends JpaRepository<Familia, Long> {
 }
