+++ src/main/java/com/Rpg/sistem_mayfair/repository/PersonagemRepository.java
@@ -3,5 +3,5 @@ package com.Rpg.sistem_mayfair.repository;
 import com.Rpg.sistem_mayfair.domain.Personagem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonagemRepository extends JpaRepository<Personagem, Integer> {
+public interface PersonagemRepository extends JpaRepository<Personagem, Long> {
 }
